TASK: support Neos 4.x (#1)
* TASK: Apply migration TYPO3.Neos.NodeTypes-201309111655

Change node type and TS object names in NodeTypes.yaml and PHP code.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.NodeTypes-20161125002300

Adjusts code to package renaming from "TYPO3.Neos.NodeTypes" to
"Neos.NodeTypes"

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.NodeTypes-20161125104800

Adjusts code to package renaming from "TYPO3.Neos.NodeTypes" to
"Neos.NodeTypes" in Fusion files

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Neos-20170115114620

Migrate usages of the Settings path
Neos.Flow.security.authentication.providers.Typo3BackendProvider to
Neos.Flow.security.authentication.providers[Neos.Neos:Backend]

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Fusion-20170120013047

Migrate name for the TypoScriptView to FusionView

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20170125103800

Migrate usages of the path [TYPO3][Flow][Security][Authentication] to
[Neos][Flow][Security][Authentication]

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Seo-20170127154600

Migrate to new namespace

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20170127183102

Migrate bootstep names.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Fusion-20180211175500

Expand Neos.Neos-FusionObjects without namespace to fully qualified
names

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Fusion-20180211184832

Remove namespace aliases and replace them with the package keys

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20180415105700

Add scalar type hint to CacheAwareInterface implementations.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* Adjust version